Partially fixed collision tunnelling

+++ Ground.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-Ground::Ground(sf::Texture*texture, float x, float y, float rectX, float rectY, float width, float height, std::string type) : Body(texture, x, y, rectX, rectY, width, height, type) {
+Ground::Ground(sf::Texture*texture, float x, float y, float rectX, float rectY, int width, int height, std::string type) : Body(texture, x, y, rectX, rectY, width, height, type) {
 
     this->m_direction = 0;
     this->m_velocityX = 0.0f;
